requirements.txt
python-telegram-bot>=21.0,<22.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.27.0,<1.0.0
pydantic>=2.8.0,<3.0.0